backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.1.8
distro==1.9.0
et_xmlfile==2.0.0
exceptiongroup==1.3.1
fastapi==0.128.8
groq==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
numpy==2.0.2
openpyxl==3.1.5
pandas==2.3.3
pydantic==2.13.2
pydantic_core==2.46.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
pytz==2026.1.post1
six==1.17.0
sniffio==1.3.1
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uvicorn==0.39.0
